refactor: move TaskId to value_objects.py
