+++ todoapp/application/task/commands/delete_task.py
@@ -5,7 +5,7 @@ from todoapp.application.common.command import Command, CommandHandler
 from todoapp.application.common.interfaces.uow import UnitOfWork
 from todoapp.application.task.exceptions import TaskAccessError
 from todoapp.application.task.interfaces.repository import TaskRepo
-from todoapp.domain.task.entities import TaskId
+from todoapp.domain.task.value_objects import TaskId
 from todoapp.domain.user.entities import UserId
