+++ todoapp/application/task/commands/update_task.py
@@ -6,7 +6,8 @@ from todoapp.application.common.interfaces.uow import UnitOfWork
 from todoapp.application.task.exceptions import TaskAccessError
 from todoapp.application.task.interfaces.repository import TaskRepo
 from todoapp.domain.common.constants import Empty
-from todoapp.domain.task.entities import Task, TaskId
+from todoapp.domain.task.entities import Task
+from todoapp.domain.task.value_objects import TaskId
 from todoapp.domain.user.entities import UserId
