+++ todoapp/domain/task/value_objects.py
@@ -1,0 +1,4 @@
+from typing import NewType
+from uuid import UUID
+
+TaskId = NewType("TaskId", UUID)
